chore:  use typescript to refactor test and fix some bugs(#42)
* style: formate code

* chore: disable no-non-null-assertion

* chore: test vHtml and vText

* chore: set optimize as true in test

* style: fix eslint warning

* chore: v-model test

* fix: v-model has not type should use vModelText

* chore: use typescript to refactor test

* fix: slots test

* feat: support isCustomElement
