+++ global.d.ts
@@ -1,2 +1,3 @@
+declare module '*.js';
 declare module '@babel/helper-module-imports';
 declare module '@babel/plugin-syntax-jsx';
