+++ packages/babel-plugin-jsx/example/index.js
@@ -1,53 +1,30 @@
-import { createApp, ref, defineComponent } from 'vue';
+import { createApp, defineComponent } from 'vue';
 
-const SuperButton = (props, context) => {
-  const obj = {
-    mouseover: () => {
-      context.emit('mouseover');
-    },
-    click: () => {
-      context.emit('click');
-    },
-  };
-  return (
-    <div class={props.class}>
-      Super
-      <button
-        on={obj}
-      >
-        { props.buttonText }
-        {context.slots.default()}
-      </button>
-    </div>
-  );
-};
-
-SuperButton.inheritAttrs = false;
-
-const App = defineComponent(() => {
-  const count = ref(0);
-  const inc = () => {
-    count.value++;
-  };
-
-  const obj = {
-    click: inc,
-    mouseover: inc,
-  };
-
-  return () => (
-    <div>
-      Foo {count.value}
-      <SuperButton
-        buttonText="VueComponent"
-        class="xxx"
-        vShow={true}
-        on={obj}
-      >
-        <button>1234</button>
-      </SuperButton>
-    </div>
-  );
+const Child = defineComponent({
+  props: ['foo'],
+  setup(props) {
+    return () => <div>{props.foo}</div>;
+  },
 });
 
-createApp(App).mount('#app');
+Child.inheritAttrs = false;
+
+const App = defineComponent({
+  data: () => ({
+    test: '1',
+  }),
+  render() {
+    return (
+      <>
+        <input type="radio" value="1" v-model={this.test} name="test" />
+        <input type="radio" value="2" v-model={this.test} name="test" />
+      </>
+    );
+  },
+});
+
+const app = createApp(App);
+
+app.mount('#app');
+
+console.log(app);
