+++ packages/babel-plugin-jsx/jest.config.js
@@ -1,3 +1,11 @@
 module.exports = {
-  setupFiles: ['./test/setup.js'],
+  setupFiles: ['./test/setup.ts'],
+  transform: {
+    '\\.(ts|tsx)$': 'ts-jest',
+  },
+  globals: {
+    'ts-jest': {
+      babelConfig: true,
+    },
+  },
 };
