+++ packages/babel-plugin-jsx/package.json
@@ -36,7 +36,9 @@
   "devDependencies": {
     "@babel/core": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
+    "@babel/preset-typescript": "^7.10.4",
     "@rollup/plugin-babel": "^5.0.3",
+    "@types/jest": "^26.0.7",
     "@types/svg-tags": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^3.6.1",
     "@typescript-eslint/parser": "^3.6.1",
@@ -47,6 +49,7 @@
     "jest": "^26.0.1",
     "regenerator-runtime": "^0.13.5",
     "rollup": "^2.13.1",
+    "ts-jest": "^26.1.3",
     "typescript": "^3.9.6",
     "vue": "3.0.0-rc.4",
     "webpack": "^4.43.0",
