+++ packages/babel-plugin-jsx/test/coverage.test.ts
@@ -1,0 +1,16 @@
+import * as fs from 'fs';
+import * as path from 'path';
+import { transformSync } from '@babel/core';
+import preset from '../babel.config.js';
+
+test('coverage', () => {
+  ['index.test.tsx', 'v-model.test.tsx']
+    .forEach((filename) => {
+      const mainTest = fs.readFileSync(path.resolve(__dirname, `./${filename}`)).toString();
+      transformSync(mainTest, {
+        babelrc: false,
+        presets: [preset],
+        filename,
+      });
+    });
+});
