+++ packages/babel-plugin-jsx/test/setup.ts
